+++ libmacho/getsegbyname.c
@@ -45,7 +45,7 @@ static type * var ## _pointer = 0
 #define SETUP_VAR(var)						\
 if ( var ## _pointer == 0) {				\
     _dyld_lookup_and_bind( STRINGIFY(_ ## var),		\
-                           (unsigned long *) & var ## _pointer, 0);	\
+                           (uint32_t *) & var ## _pointer, 0);	\
 }
 #define USE_VAR(var) (* var ## _pointer)
 #endif
@@ -66,7 +66,7 @@ char *segname)
 {
     static struct mach_header *mhp = NULL;
     struct segment_command *sgp;
-    unsigned long i;
+    uint32_t i;
 #ifndef RLD
 #ifndef __OPENSTEP__
 	if(mhp == NULL)
@@ -99,7 +99,7 @@ char *segname)
 {
     static struct mach_header_64 *mhp = NULL;
     struct segment_command_64 *sgp;
-    unsigned long i;
+    uint32_t i;
 
 	if(mhp == NULL)
 #ifndef RLD
