+++ otool/otool.h
@@ -52,8 +52,8 @@ extern char *pflag; 	/* procedure name to start disassembling from */
 extern char *segname,
      *sectname;	    /* name of the section to print the contents of */
 
-unsigned long m68k_usrstack(void);
-unsigned long m88k_usrstack(void);
-unsigned long i386_usrstack(void);
-unsigned long hppa_usrstack(void);
-unsigned long sparc_usrstack(void);
+uint32_t m68k_usrstack(void);
+uint32_t m88k_usrstack(void);
+uint32_t i386_usrstack(void);
+uint32_t hppa_usrstack(void);
+uint32_t sparc_usrstack(void);
